refactor: move common to common.error
